requirements.txt
asgiref==3.4.1
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-mptt==0.12.0
django==3.2.5
gunicorn==20.1.0
pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
django-tinymce==3.4.0
django-simple-captcha==0.5.17
cloudinary==1.29.0
git+https://github.com/Daniil-7/python-sdk.git



django-heroku==0.3.1
